% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{DATE_ORDER_SEQUENCE}
\alias{DATE_ORDER_SEQUENCE}
\title{Expected chronological order of cascade milestone dates (Descriptive
Study Plan SS6, point 3)}
\usage{
DATE_ORDER_SEQUENCE
}
\description{
Expected chronological order of cascade milestone dates (Descriptive
Study Plan SS6, point 3)
}
