% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{.branch_factor}
\alias{.branch_factor}
\title{Map a set of mutually-exclusive 0/1 columns to a single labelled factor}
\usage{
.branch_factor(df, column_labels)
}
\arguments{
\item{df}{A data frame.}

\item{column_labels}{A named character vector: names are columns of
\code{df}, values are the factor label to use when that column is \code{1}.}
}
\value{
A factor, \code{NA} where none of the columns is \code{1}.
}
\description{
Map a set of mutually-exclusive 0/1 columns to a single labelled factor
}
\keyword{internal}
