% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{.suppression_caption}
\alias{.suppression_caption}
\title{Standard disclosure-control caption for suppressed-cell charts}
\usage{
.suppression_caption()
}
\description{
Standard disclosure-control caption for suppressed-cell charts
}
\keyword{internal}
