% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{N_ROWS_EXPECTED}
\alias{N_ROWS_EXPECTED}
\alias{N_COLS_EXPECTED}
\title{Expected shape of the raw dataset (regression guard)}
\usage{
N_ROWS_EXPECTED

N_COLS_EXPECTED
}
\description{
Expected shape of the raw dataset (regression guard)
}
