% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{SMALL_CELL_THRESHOLD}
\alias{SMALL_CELL_THRESHOLD}
\title{Small-cell suppression threshold (Descriptive Study Plan SS11)}
\usage{
SMALL_CELL_THRESHOLD
}
\description{
Small-cell suppression threshold (Descriptive Study Plan SS11)
}
